% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummymodel.R
\name{baseshape}
\alias{baseshape}
\title{baseshape generator function
generating a line, trangle or square}
\usage{
baseshape(points = 2)
}
\arguments{
\item{points}{an integer that is either 2, 3, or 4 to determine init number points}
}
\value{
a line, triangle or square depend on input
}
\description{
baseshape generator function
generating a line, trangle or square
}
