% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummymodel.R
\name{baseshapenew}
\alias{baseshapenew}
\title{the new surface generator function}
\usage{
baseshapenew(controlpoints)
}
\arguments{
\item{controlpoints}{the n * 3 matrix that lay on the same plane}
}
\value{
controlpoints  same as input if no exception exist
surface        an 1 * n matrix or n len vector that represent which points create surface.
}
\description{
the new surface generator function
}
