% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummymodel.R
\name{join}
\alias{join}
\title{join model function}
\usage{
join(points1, surface1, points2, surface2)
}
\arguments{
\item{points1}{the model1 control points}

\item{surface1}{the model1 surface}

\item{points2}{the model2 control points}

\item{surface2}{the model2 surface}
}
\value{
points the combined control points
surface the combined surface
}
\description{
join model function
}
