% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummymodel.R
\name{sectionview}
\alias{sectionview}
\title{Title}
\usage{
sectionview(points, surface, plane = c(0, 0, 1), k = 0)
}
\arguments{
\item{points}{the P * 3 matrix that is generated by other base functions}

\item{surface}{the N * 3 matrix indicate the surface component}

\item{plane}{the plane constant variable x y z store as c(x,y,z) that we want to cut though the model. default is c(0,0,1) or xy plane}

\item{k}{the scaler function that make the plane shift. equations is x + y + z = k default k = 0}
}
\value{
the function do not return anything but draw a graph
}
\description{
Title
}
