% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummymodel.R
\name{spin}
\alias{spin}
\title{spin function
rotate model at (0,0,0) axis with provided direction}
\usage{
spin(points, direction = 0, angle = pi/2)
}
\arguments{
\item{points}{the P * 3 matrix that is generated by other base functions}

\item{direction}{an integer that is either 0, 1, 2 corresponding xy, xz, yz plane we perform rotate}

\item{angle}{how much we rotate the graph in radian, default is half pi radian or 90 degree}
}
\value{
newpoints   the P * 3 matrix of points after the rotation is performed.
}
\description{
spin function
rotate model at (0,0,0) axis with provided direction
}
