% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummymodel.R
\name{baseshapenew}
\alias{baseshapenew}
\title{the new surface generator function}
\usage{
baseshapenew(preset = NULL, controlpoints = NULL)
}
\arguments{
\item{preset}{a scaler that will pull preset model. 1 = square,}

\item{controlpoints}{the 3 * 3 matrix that lay on the same plane}
}
\value{
controlpoints  same as input if no exception exist
surface        an n * 3 matrix  that represent which points create triangle surface.
}
\description{
the new surface generator function
}
