% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummymodel.R
\name{simplifymodel}
\alias{simplifymodel}
\title{simplify model function
combining duplicated or close control points}
\usage{
simplifymodel(points, connections)
}
\arguments{
\item{points}{the P * 3 matrix that is generated by other base functions}

\item{connections}{the P * P matrix indicate the connections between points}
}
\value{
points    the (P - x) * 3 matrix, combined x pair points in simplifying procedure
connections   the (P - x) * (P - x) matrix indicate the connections between new points
}
\description{
simplify model function
combining duplicated or close control points
}
